requirements.txt
streamlit
google.generativeai
datetime